+++ public/styles/inputs.styl
@@ -70,8 +70,6 @@ input
   background: none
   resize: none
   width: 100%
-  label&
-    padding: $input-padding
     
   &:empty:before
     content: attr(placeholder)
@@ -80,7 +78,14 @@ input
 
   &.error
     border-color: warning
+
+.content-editable-prefix-container
+  display: flex
   
+.content-editable-prefix
+    padding: $input-padding
+    width: auto
+    
 .editable-field-container
   position:relative
